Sephora_BE/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
python-decouple
django-cors-headers
firebase-admin
